+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/ConfluentSchemaRegistryProvider.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.engine.avro.schemaregistry.confluent
 
 import cats.data.Validated.{Invalid, Valid}
 import cats.data.{NonEmptyList, ValidatedNel}
+import io.circe.Json
 import org.apache.avro.Schema
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{CachedConfluentSchemaRegistryClientFactory, ConfluentSchemaRegistryClientFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.formatter.ConfluentAvroToJsonFormatterFactory
@@ -10,13 +11,12 @@ import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.{C
 import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaRegistryError, SchemaRegistryProvider, SchemaRegistryUnsupportedTypeError}
 import pl.touk.nussknacker.engine.avro.serialization.{KafkaAvroDeserializationSchemaFactory, KafkaAvroSerializationSchemaFactory}
 import pl.touk.nussknacker.engine.kafka.RecordFormatterFactory
+import pl.touk.nussknacker.engine.kafka.consumerrecord.ConsumerRecordToJsonFormatterFactory
 
 class ConfluentSchemaRegistryProvider(val schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
                                       val serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
-                                      val deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory) extends SchemaRegistryProvider {
-
-  override def recordFormatterFactory: RecordFormatterFactory =
-    new ConfluentAvroToJsonFormatterFactory(schemaRegistryClientFactory)
+                                      val deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
+                                      val recordFormatterFactory: RecordFormatterFactory) extends SchemaRegistryProvider {
 
   override def validateSchema(schema: Schema): ValidatedNel[SchemaRegistryError, Schema] =
   /* kafka-avro-serializer does not support Array at top level
@@ -42,17 +42,20 @@ object ConfluentSchemaRegistryProvider extends Serializable {
     ConfluentSchemaRegistryProvider(
       schemaRegistryClientFactory,
       new ConfluentAvroSerializationSchemaFactory(schemaRegistryClientFactory),
-      new ConfluentKeyValueKafkaAvroDeserializationFactory(schemaRegistryClientFactory)
+      new ConfluentKeyValueKafkaAvroDeserializationFactory(schemaRegistryClientFactory),
+      new ConfluentAvroToJsonFormatterFactory(schemaRegistryClientFactory)
     )
   }
 
   def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
             serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
-            deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory): ConfluentSchemaRegistryProvider = {
+            deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
+            recordFormatterFactory: RecordFormatterFactory): ConfluentSchemaRegistryProvider = {
     new ConfluentSchemaRegistryProvider(
       schemaRegistryClientFactory,
       serializationSchemaFactory,
-      deserializationSchemaFactory
+      deserializationSchemaFactory,
+      recordFormatterFactory
     )
   }
 
@@ -60,7 +63,8 @@ object ConfluentSchemaRegistryProvider extends Serializable {
     ConfluentSchemaRegistryProvider(
       schemaRegistryClientFactory,
       new ConfluentJsonPayloadSerializerFactory(schemaRegistryClientFactory),
-      new ConfluentKeyValueKafkaJsonDeserializerFactory(schemaRegistryClientFactory)
+      new ConfluentKeyValueKafkaJsonDeserializerFactory(schemaRegistryClientFactory),
+      new ConsumerRecordToJsonFormatterFactory[Json, Json]
     )
   }